Add more clear error for logical-choices bad usage
Add a more clear error for bad usage choices option when coupled with
logical options.

Why:

When adding a logical option to the CLI it is not possible to prevent
the user to use the *choices* optional argument, that is intended to be
used for limiting the range of admissible values of the passed option.
However, logical type is be definition limited to only 2 values, namely
*.true.* or *.false.*, thus it is meaningless to use the *choices*
optional argument with logical option.

This change addresses the need by:

Add a more clear error message when the user try to retry the value of
a logical option for which a *choices* set has been defined.

Side effects:

Nothing.

Unrelated modifications:

Add the *lib* and *tests* directories inside *src* one for a better
separation of the library sources and the tests suite ones. The fobos
file has been modified accordingly.