Fix bug #71 group default value for not invoked group
Short description

Querying the default values of CLAs belonging to not invoked group
returns unpredictable results instead of the default value set
initializing the CLI.

Why:

Sane fallback.

This change addresses the need by:

Eliminating the check if the group has been invoked or not.

Side effects:

Nothing, but a warning could be like to be added in the case a not
invoked group is queried.