Fix issues #94 #95
Fix issues #94 #95: switch from Travis CI to github actions for
performing CI and revert back the creation of release tagged version
with automatic assets upload.

Trim the verion number from the name of the tarball of the progect: the
install script should be now working.

The release assets is uploaded always with the last tag and on the
master branch.